refactor: move tarp plotting to analysis, fix tutorial.
